pyproject.toml
[project]
name = "Crafter FOSS"
version = "0.0.3"
description = "Craft BootAnimations for Android devices FOSS."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "daedsky", email = "you@example.com" }
]
dependencies = [
    "flet", "pillow", "flet-permission-handler", "pyjnius"
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "io.crafter"
build_number = 3
compile.app = true # --compile-app
compile.packages = true # --compile-packages
compile.cleanup = true # --cleanup-on-compile

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "Crafter FOSS"

# company name to display in about app dialogs
company = "Fusion Reactors"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by daedsky"

[tool.flet.app]
path = "src"

[tool.flet.android]
split_per_abi = true
min_sdk_version = 21
target_sdk_version = 34

[tool.flet.app.boot_screen]
show = true
message = "Preparing the app for its first launch…"

[tool.flet.app.startup_screen]
show = true
message = "Starting up the app…"

[tool.flet.android.permission]
#"android.permission.STORAGE" = true
"android.permission.MANAGE_EXTERNAL_STORAGE" = true
